Revert "added filters"
