+++ source/_patterns/02-components/05-search/search.mustache
@@ -1,6 +1,6 @@
 {{#search}}
     <div class="search-form">
-        <form class="js-search-form">
+        <form class="js-search-form" data-api="../../api/doctors.json">
             <fieldset class="js-search-fieldset">
                 <div class="search">
                     <div class="search__input">
